feat(pocha-manage): move PochaForm body into PochaFormDialog, delete legacy PochaForm (lane 2.11b)
Task 2 of lane 2.11b PochaForm UX redesign:
- Move full form state, cross-field validation, and onSubmit handler from
  PochaForm.tsx into PochaFormDialog.tsx
- Layout: DialogTitle | scrollable Form body (PochaInfoFields + Divider +
  PochaMenuFields) | sticky DialogFooter with cancel + submit buttons
- Cross-field error Alert relocated to DialogFooter (above the action row,
  spans full width)
- Auto-close dialog after successful create/update (after SWR mutates)
- Cancel button closes dialog without submitting
- Delete legacy PochaForm.tsx — only consumer (page.tsx) was already swapped
  to PochaFormDialog in Task 1
- Relies on DS web 1.0.12+: DialogContent now ships max-h-full + flex-col so
  no consumer-side max-h override needed (DS commit 380115d, will land in
  next DS bump)
